modifies tests so that the one that's failing for a different reason is separate from the passing ones. Also added more tests
